[ClockPicker] Move getHours,getMinutes to the module where it's used
