+++ packages/material-ui-lab/src/internal/pickers/time-utils.ts
@@ -1,5 +1,4 @@
 import { ParsableDate } from './constants/prop-types';
-import { CLOCK_WIDTH } from './constants/dimensions';
 import { MuiPickersAdapter } from './hooks/useUtils';
 
 type Meridiem = 'am' | 'pm' | null;
@@ -31,62 +30,6 @@ export const convertToMeridiem = <TDate>(
 ) => {
   const newHoursAmount = convertValueToMeridiem(utils.getHours(time), meridiem, ampm);
   return utils.setHours(time, newHoursAmount);
-};
-
-const clockCenter = {
-  x: CLOCK_WIDTH / 2,
-  y: CLOCK_WIDTH / 2,
-};
-
-const baseClockPoint = {
-  x: clockCenter.x,
-  y: 0,
-};
-
-const cx = baseClockPoint.x - clockCenter.x;
-const cy = baseClockPoint.y - clockCenter.y;
-
-const rad2deg = (rad: number) => rad * 57.29577951308232;
-
-const getAngleValue = (step: number, offsetX: number, offsetY: number) => {
-  const x = offsetX - clockCenter.x;
-  const y = offsetY - clockCenter.y;
-
-  const atan = Math.atan2(cx, cy) - Math.atan2(x, y);
-
-  let deg = rad2deg(atan);
-  deg = Math.round(deg / step) * step;
-  deg %= 360;
-
-  const value = Math.floor(deg / step) || 0;
-  const delta = x ** 2 + y ** 2;
-  const distance = Math.sqrt(delta);
-
-  return { value, distance };
-};
-
-export const getMinutes = (offsetX: number, offsetY: number, step = 1) => {
-  const angleStep = step * 6;
-  let { value } = getAngleValue(angleStep, offsetX, offsetY);
-  value = (value * step) % 60;
-
-  return value;
-};
-
-export const getHours = (offsetX: number, offsetY: number, ampm: boolean) => {
-  const { value, distance } = getAngleValue(30, offsetX, offsetY);
-  let hour = value || 12;
-
-  if (!ampm) {
-    if (distance < 90) {
-      hour += 12;
-      hour %= 24;
-    }
-  } else {
-    hour %= 12;
-  }
-
-  return hour;
 };
 
 export function getSecondsInDay(date: unknown, utils: MuiPickersAdapter) {
